RDFLib Syntax Check
Using RDFLib, a Python library designed for working with RDF (Resource Description Framework) data, to detect syntax errors in the generated Turtle ontology. 

When you parse a Turtle file or string using RDFLib, if the Turtle syntax is incorrect, RDFLib will raise a syntax error exception, indicating that there's a problem with the Turtle syntax.